decoupled the employee service from the calculator itself
